+++ test/unit/io/views/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_subdirectories()
+
+seqan3_test(async_input_buffer_test.cpp)
